+++ R/chop.R
@@ -320,6 +320,12 @@ chop_deciles <- function(x, ...) {
 #' @inheritParams chop
 #' @inherit chop-doc params return
 #'
+#' @details
+#' `chop_equally()` uses [chop_quantiles()] under the hood. If `x` has duplicate
+#' elements, you are not guaranteed to get `groups`. See the examples.
+#'
+#'
+#'
 #' @family chopping functions
 #'
 #' @export
@@ -327,6 +333,10 @@ chop_deciles <- function(x, ...) {
 #' @examples
 #' chop_equally(1:10, 5)
 #'
+#' # You can't always guarantee `groups` groups:
+#' dupes <- c(1, 1, 1, 2, 3, 4, 4, 4)
+#' quantile(dupes, 0:4/4)
+#' chop_equally(dupes, 4)
 chop_equally <- function (
                   x,
                   groups,